+++ src/screens/validators/components/list/components/validators/components/mobile/index.tsx
@@ -6,8 +6,7 @@ import AutoSizer from 'react-virtualized-auto-sizer';
 import { Divider } from '@material-ui/core';
 import { AvatarName } from '@components';
 import {
-  NODE_DETAILS,
-  VALIDATOR_DETAILS,
+  VALIDATOR_DETAILS, NODE_DETAILS,
 } from '@utils/go_to_page';
 import {
   useList,
@@ -40,7 +39,7 @@ const Mobile: React.FC<{
           address={x.validator.address}
           imageUrl={x.validator.imageUrl}
           name={x.validator.name}
-          href={x.validator.address ? VALIDATOR_DETAILS : NODE_DETAILS}
+          href={x.isNode ? NODE_DETAILS : VALIDATOR_DETAILS}
         />
       ),
       locked: (
@@ -50,9 +49,11 @@ const Mobile: React.FC<{
           content={formatNumber(x.locked.value, 2)}
         />
       ),
-      topUp: `${formatNumber(x.topUp.value, 2)} ${x.stake.displayDenom.toUpperCase()}`,
-      stake: `${formatNumber(x.stake.value, 2)} ${x.stake.displayDenom.toUpperCase()}`,
-      nodes: numeral(x.nodes).format('0,0'),
+      stake: `${formatNumber(x.stake.value, x.stake.exponent)} ${x.stake.displayDenom.toUpperCase()}`,
+      nodes: x.nodes ? numeral(x.nodes).format('0,0') : '-',
+      delegators: x.delegators ? numeral(x.delegators).format('0,0') : '-',
+      commission: x.commission ? `${numeral(x.commission * 100).format('0,0.[00]')}%` : '-',
+      apr: x.apr ? `${x.apr}%` : '-',
     });
   });