+++ src/screens/validators/components/list/types.ts
@@ -6,20 +6,14 @@ export type TabType = {
 
 export type ValidatorType = {
   validator: AvatarName;
-  stake: TokenUnit;
   locked: TokenUnit;
-  topUp: TokenUnit;
-  stakePercent: number;
-  nodes: number;
-}
-
-export type ProviderType = {
-  validator: AvatarName;
   stake: TokenUnit;
-  nodes: number;
+  stakePercent: number;
   commission: number;
+  nodes: number;
   apr: number;
   delegators: number;
+  isNode: boolean;
 }
 
 export type SearchType = string;
@@ -30,5 +24,4 @@ export type ValidatorsState = {
   tab: number;
   search: SearchType;
   validators: ValidatorType[];
-  providers: ProviderType[];
 }