+++ CHANGELOG.md
@@ -20,3 +20,4 @@
 - Update homepage to setup some loading indicator
 - Remove tx options from settings
 - Add value to tx details operations with link to token details
+- Merged validator tabs together